Refactor updater test API to test common pkg (#4626)
* Refactor updater test API to test common pkg
* Fix up imports for updates test package
