+++ tests/e2e/update/apiconversion/apiconversion_test.go
@@ -5,13 +5,14 @@ package apiconversion
 
 import (
 	"fmt"
+	"time"
+
 	. "github.com/onsi/ginkgo/v2"
 	"github.com/verrazzano/verrazzano/pkg/constants"
 	"github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1beta1"
 	"github.com/verrazzano/verrazzano/tests/e2e/pkg"
 	"github.com/verrazzano/verrazzano/tests/e2e/pkg/test/framework"
-	"github.com/verrazzano/verrazzano/tests/e2e/update"
-	"time"
+	"github.com/verrazzano/verrazzano/tests/e2e/pkg/update"
 )
 
 const (
