+++ tests/e2e/update/availability/availability_suite_test.go
@@ -4,9 +4,10 @@
 package availability
 
 import (
+	"testing"
+
 	"github.com/onsi/ginkgo/v2"
 	"github.com/onsi/gomega"
-	"testing"
 )
 
 func TestAvailability(t *testing.T) {
