+++ tests/e2e/update/fluentd/fluentd_update_test.go
@@ -5,6 +5,8 @@ package fluentd
 
 import (
 	"fmt"
+	"time"
+
 	"github.com/google/uuid"
 	. "github.com/onsi/ginkgo/v2"
 	"github.com/onsi/gomega"
@@ -13,7 +15,6 @@ import (
 	pcons "github.com/verrazzano/verrazzano/platform-operator/constants"
 	"github.com/verrazzano/verrazzano/tests/e2e/pkg"
 	"github.com/verrazzano/verrazzano/tests/e2e/pkg/test/framework"
-	"time"
 )
 
 const (
