+++ tests/e2e/update/fluentdextes/fluentdextes_test.go
@@ -16,7 +16,7 @@ import (
 	"github.com/verrazzano/verrazzano/tests/e2e/multicluster"
 	"github.com/verrazzano/verrazzano/tests/e2e/pkg"
 	"github.com/verrazzano/verrazzano/tests/e2e/pkg/test/framework"
-	"github.com/verrazzano/verrazzano/tests/e2e/update"
+	"github.com/verrazzano/verrazzano/tests/e2e/pkg/update"
 	"github.com/verrazzano/verrazzano/tests/e2e/update/fluentd"
 	corev1 "k8s.io/api/core/v1"
 )
