+++ tests/e2e/update/jaeger/jaeger_update_test.go
@@ -4,15 +4,16 @@
 package jaeger
 
 import (
+	"strings"
+	"time"
+
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	"github.com/verrazzano/verrazzano/pkg/k8sutil"
 	"github.com/verrazzano/verrazzano/platform-operator/constants"
 	"github.com/verrazzano/verrazzano/tests/e2e/pkg"
 	"github.com/verrazzano/verrazzano/tests/e2e/pkg/test/framework"
-	"github.com/verrazzano/verrazzano/tests/e2e/update"
-	"strings"
-	"time"
+	"github.com/verrazzano/verrazzano/tests/e2e/pkg/update"
 )
 
 const (
